+++ examples/point/scatter/demo/scatter-shapes.ts
@@ -1,6 +1,6 @@
 import { Chart } from '@antv/g2';
 
-fetch('../data/scatter.json')
+fetch('https://gw.alipayobjects.com/os/antvdemo/assets/data/scatter.json')
   .then((res) => res.json())
   .then((data) => {
     const chart = new Chart({