+++ src/chart/controller/tooltip.ts
@@ -37,7 +37,6 @@ export default class Tooltip extends Controller<TooltipOption> {
   private guideGroup: IGroup;
 
   private isLocked: boolean = false;
-  private isVisible: boolean = true;
   private items;
   private title: string;
   private point: Point;
@@ -48,10 +47,12 @@ export default class Tooltip extends Controller<TooltipOption> {
 
   public init() {}
 
-  public render() {
+  private isVisible() {
     const option = this.view.getOptions().tooltip;
-    this.isVisible = option !== false;
+    return option !== false;
   }
+
+  public render() {}
 
   /**
    * Shows tooltip
@@ -59,7 +60,7 @@ export default class Tooltip extends Controller<TooltipOption> {
    */
   public showTooltip(point: Point) {
     this.point = point;
-    if (!this.isVisible) {
+    if (!this.isVisible()) {
       // 如果设置 tooltip(false) 则始终不显示
       return;
     }
@@ -322,6 +323,17 @@ export default class Tooltip extends Controller<TooltipOption> {
   public update() {
     if (this.point) {
       this.showTooltip(this.point);
+    }
+    
+    if (this.tooltip) {
+      // #2279 修复resize之后tooltip越界的问题
+      // 确保tooltip已经创建的情况下
+      const canvas = this.view.getCanvas();
+      // 更新 region
+      this.tooltip.set('region', {
+        start: { x: 0, y: 0 },
+        end: { x: canvas.get('width'), y: canvas.get('height') },
+      });
     }
   }